Escape single quotes in response descriptions
If response descriptions have a single quote - such as "don't" -
then the description of the "should have" test for this produces
an invalid JavaScript string.  This escapes these single quotes
where they occur in response descriptions.

Written by someone completely unused to writing JavaScript and who
doesn't use Node, so I'm sorry if it's not idiomatic.

Signed-off-by: Paul Wayper <paulway@redhat.com>